requirements.txt
Django==4.1.7
djangorestframework==3.14.0
mysqlclient
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
Pillow==9.5.0
gunicorn==20.1.0
django-otp<4.4.0
drf-yasg==1.21.7
